turns out nvidia needs highp???

+++ shader/background.frag
@@ -1,6 +1,7 @@
 #version 320 es
 
-precision mediump float;
+//precision mediump float;
+precision highp float;
 
 uniform float time;
 uniform vec2 resolution;
